Various changes...
Changed axis scaling to only scale to fit x-axis and set the y-axis to be the same size so that the ball curve will be correct. This does however create the possibility that the ball can go out of bounds at the top of the screen.

Also added one decimal to axis

Also made it possible to update with a different timestep than delta